+++ app/assets/javascripts/angular/components/organizationContactImport/organizationContactImportStep4.component.js
@@ -1,5 +1,5 @@
 import template from './organizationContactImportStep4.html';
-import checkIcon from '../../../../images/icon-check.svg';
+import checkIcon from '../../../../images/icons/icon-check.svg';
 
 angular.module('missionhubApp').component('organizationContactImportStep4', {
     require: {
